huge update

+++ commands/say.js
@@ -1,8 +1,11 @@
 exports.run = (client, message, args) => {
-  const ana = " - Anonymous"
-  let text = args.join(" " + ana);
+if (message.author.id == process.env.SuperUserID){
+  let text = args.join(" ");
   message.delete();
   message.channel.send(text);
+  }else{
+  return;
+  }
 }
 
 exports.name = "say";
